게시물 이미지 병합 수정

+++ insider/src/main/java/com/kh/insider/repo/BoardRepo.java
@@ -8,7 +8,8 @@ import com.kh.insider.vo.BoardListVO;
 public interface BoardRepo {
 	List<BoardDto> selectListPaging(int page);
 
-	int sequence();
+	Long sequence();
+	
 	void insert(BoardDto boardDto);
 
 	void updateLikeCount(int boardNo, int count);
